+++ classes/Admin/AdminController.php
@@ -528,7 +528,7 @@ class AdminController
             // Base64 decode the route
             $data['route'] = isset($data['route']) ? base64_decode($data['route']) : null;
             $data['filters'] = json_decode($options['filters'] ?? '{}', true) + ['type' => ['dir']];
-            $data['language'] = $this->getLanguage();
+            $data['lang'] = $this->getLanguage();
 
             $initial = $data['initial'] ?? null;
             if ($initial) {